requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.25.0
crcmod==1.7
cryptography==40.0.1
datasets==2.8.0
dill==0.3.6
einops==0.6.0
filelock==3.10.7
flatbuffers==23.3.3
frozenlist==1.3.3
fsspec==2023.3.0
ftfy==6.1.1
gast==0.4.0
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
huggingface-hub==0.13.3
idna==3.4
jax==0.4.7
Jinja2==3.1.2
jmespath==0.10.0
jsonplus==0.8.0
keras==2.12.0
libclang==16.0.0
lightning-utilities==0.8.0
lit==15.0.7
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.0.4
modelscope==1.4.2
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.0
numpy==1.23.5
oauthlib==3.2.2
open-clip-torch==2.16.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
oss2==2.17.0
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.2.0
protobuf==3.20.3
py-cord==2.4.1
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
python-dateutil==2.8.2
python-dotenv==1.0.0
pytorch-lightning==2.0.0
pytz==2023.2
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
scipy==1.10.1
sentencepiece==0.1.97
simplejson==3.18.4
six==1.16.0
sortedcontainers==2.4.0
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
timm==0.6.13
tomli==2.0.1
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchmetrics==0.11.4
torchvision==0.15.1+cu117
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.14.1
xxhash==3.2.0
yapf==0.32.0
yarl==1.8.2
